+++ activemodel/lib/active_model/gem_version.rb
@@ -9,7 +9,7 @@ module ActiveModel
   module VERSION
     MAJOR = 7
     MINOR = 0
-    TINY  = 1
+    TINY  = 2
     PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")